Use a central iteration type
