+++ app/javascript/components/track/IterationSummary.tsx
@@ -1,66 +1,14 @@
 import React, { useEffect, useState, useRef } from 'react'
-import { SubmissionTestsStatus } from '../editor/types'
 import { fromNow } from '../../utils/time'
 import { SubmissionMethodIcon } from './iteration-summary/SubmissionMethodIcon'
 import { ProcessingStatusSummary } from './iteration-summary/ProcessingStatusSummary'
 import { AnalysisStatusSummary } from './iteration-summary/AnalysisStatusSummary'
 import { IterationChannel } from '../../channels/iterationChannel'
-
-export type Iteration = {
-  uuid: string
-  idx: number
-  status: IterationStatus
-  numEssentialAutomatedComments: number
-  numActionableAutomatedComments: number
-  numNonActionableAutomatedComments: number
-  submissionMethod: SubmissionMethod
-  createdAt: Date
-  testsStatus: SubmissionTestsStatus
-  links: IterationLinks
-}
-
-type IterationLinks = {
-  self: string
-}
-
-export enum IterationStatus {
-  TESTING = 'testing',
-  TESTS_FAILED = 'tests_failed',
-  ANALYZING = 'analyzing',
-  ESSENTIAL_AUTOMATED_FEEDBACK = 'essential_automated_feedback',
-  ACTIONABLE_AUTOMATED_FEEDBACK = 'actionable_automated_feedback',
-  NON_ACTIONABLE_AUTOMATED_FEEDBACK = 'non_actionable_automated_feedback',
-  NO_AUTOMATED_FEEDBACK = 'no_automated_feedback',
-}
-
-export enum SubmissionMethod {
-  CLI = 'cli',
-  API = 'api',
-}
-
-export enum RepresentationStatus {
-  NOT_QUEUED = 'not_queued',
-  QUEUED = 'queued',
-  APPROVED = 'approved',
-  DISAPPROVED = 'disapproved',
-  INCONCLUSIVE = 'inconclusive',
-  EXCEPTIONED = 'exceptioned',
-  CANCELLED = 'cancelled',
-}
-
-export enum AnalysisStatus {
-  NOT_QUEUED = 'not_queued',
-  QUEUED = 'queued',
-  APPROVED = 'approved',
-  DISAPPROVED = 'disapproved',
-  INCONCLUSIVE = 'inconclusive',
-  EXCEPTIONED = 'exceptioned',
-  CANCELLED = 'cancelled',
-}
+import { Iteration } from '../types'
 
 const SUBMISSION_METHOD_LABELS = {
-  [SubmissionMethod.CLI]: 'CLI',
-  [SubmissionMethod.API]: 'API',
+  cli: 'CLI',
+  api: 'API',
 }
 
 export function IterationSummary(props: {
